scripts/requirements.txt
psycopg2-binary
boto3
pytest
unittest